refactor firebase functions code
