+++ firebase_functions/functions/package.json
@@ -5,6 +5,7 @@
     "build": "tsc",
     "serve": "npm run build && firebase emulators:start --only functions,firestore,ui",
     "shell": "npm run build && firebase functions:shell",
+    "watch": "tsc --watch",
     "start": "npm run shell",
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log"
@@ -14,11 +15,13 @@
   },
   "main": "lib/index.js",
   "dependencies": {
+    "express": "^4.17.3",
     "firebase-admin": "^9.8.0",
     "firebase-functions": "^3.14.1",
     "http-status-codes": "^2.2.0"
   },
   "devDependencies": {
+    "@types/express": "^4.17.13",
     "@typescript-eslint/eslint-plugin": "^3.9.1",
     "@typescript-eslint/parser": "^3.8.0",
     "eslint": "^7.6.0",
